feat: add Inventory service

+++ src/main/java/com/krystianwoloszun/inv360/warehouse/WarehouseService.java
@@ -46,6 +46,9 @@ public class WarehouseService {
 
     public void deleteWarehouse(Long id) {
         Warehouse existingWarehouse = getWarehouseById(id);
+        if (existingWarehouse == null) {
+            throw new WarehouseNotFoundException("Warehouse with ID " + id + " not found.");
+        }
         warehouseRepository.delete(existingWarehouse);
     }
